Moves scss files
