+++ src/assets/scss/antd-overrides.scss
