+++ src/assets/scss/responsive-carousel-override.scss
