+++ src/components/shared/ImageUploadAndNameInputs.js
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react';
 import { Layout, Row, Col, Input, Form, Upload, message } from 'antd';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faCloudUploadAlt } from '@fortawesome/free-solid-svg-icons';
-import 'scss/antd-overrides.scss';
+import 'assets/scss/antd-overrides.scss';
 
 function getBase64(image, callback) {
   const reader = new FileReader();
